potato diease/backend/requirements.txt
fastapi
uvicorn
langchain
google-generativeai
python-multipart
Pillow
python-dotenv
Flask==2.4.0
flask-cors==3.0.11
python-dotenv==1.0.0
google-generativeai==0.2.0
langchain==0.0.303
Pillow==10.0.0
